[AMQ-8354] extract queue sinchronization to it's own class.
